requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.2
astroid==2.11.5
async-timeout==3.0.1
attrs==22.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.6.0
awsebcli==3.20.6
beautifulsoup4==4.9.3
billiard==3.6.4.0
black==23.1.0
botocore==1.29.99
branca==0.5.0
CacheControl==0.12.10
cachetools==5.0.0
celery==5.2.7
cement==2.8.2
certifi==2022.12.7
cffi==1.15.1
channels==3.0.4
channels-redis==3.3.0
chardet==4.0.0
charset-normalizer==3.0.1
clear-screen==0.1.14
click==8.1.3
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
cssutils==2.2.0
daphne==3.0.2
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.5.1
distlib==0.3.8
dj-database-url==1.0.0
Django==4.1.3
django-allauth==0.44.0
django-bootstrap5==22.1
django-celery-beat==2.4.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
django-debug-toolbar==4.1.0
django-dotenv==1.4.2
django-environ==0.4.5
django-extensions==3.2.1
django-filter==22.1
django-heroku==0.3.1
django-import-export==2.5.0
django-inlinecss==0.3.0
django-model-utils==4.2.0
django-phonenumber-field==7.0.0
django-redis==5.2.0
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-rest-passwordreset==1.1.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.1.0
drf-spectacular==0.25.1
drf-yasg==1.21.4
et-xmlfile==1.1.0
ez-setup==0.9
facebook-sdk==3.1.0
factory-boy==3.2.1
Faker==15.3.1
fb==0.4.0
fcm-django==1.0.8
filelock==3.13.4
firebase-admin==5.2.0
folium==0.12.1.post1
frappe-bench==5.22.3
future==0.18.2
geocoder==1.38.1
gevent==21.12.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.2.1
googleapis-common-protos==1.54.0
greenlet==1.1.2
grequests==0.6.0
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.1.0
hiredis==2.0.0
honcho==1.1.0
httpie==3.2.1
httplib2==0.20.4
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.17.3
kombu==5.2.4
lazy-object-proxy==1.7.1
lxml==4.9.1
markdown-it-py==2.2.0
MarkupPy==1.14
MarkupSafe==2.1.1
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
netifaces==0.11.0
numpy==1.26.0
oauthlib==3.1.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.7
packaging==23.0
pandas==1.4.2
pathspec==0.11.0
phonenumbers==8.13.0
Pillow==9.3.0
pipenv==2023.12.1
platformdirs==4.2.0
prompt-toolkit==3.0.19
proto-plus==1.20.1
protobuf==3.19.4
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.16.0
pygame==2.5.2
Pygments==2.14.0
PyJWT==2.6.0
pylint==2.13.9
pynliner==0.8.0
pyOpenSSL==20.0.1
pyparsing==3.0.9
pypiwin32==223
pyrsistent==0.19.3
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
pywin32==306
PyYAML==6.0
ratelim==0.1.6
recordtype==1.4
redis==3.5.3
requests==2.28.2
requests-oauthlib==1.3.0
requests-toolbelt==0.10.1
rich==13.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
semantic-version==2.8.5
service-identity==21.1.0
setuptools-scm==7.1.0
simplejson==3.17.6
six==1.16.0
smmap==5.0.1
soupsieve==2.2
South==1.0.2
sqlparse==0.4.3
tablib==3.0.0
termcolor==1.1.0
toml==0.10.2
tomli==2.0.1
twilio==7.15.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
twitter==1.19.3
txaio==21.2.1
typing_extensions==4.4.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.14
urlparse2==1.1.1
vine==5.0.0
virtualenv==20.25.3
wcwidth==0.1.7
whitenoise==6.2.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
